requirements.txt
networkx
openai
numpy
python-dotenv
streamlit
flask
pyvis
requests
streamlit-chat
streamlit-autorefresh
pandas
docx
werkzeug